cleaned up the radiation.py mess, moved some functions into other files like albedo and radiation_ipw fix issue #156
