+++ examples/SMRF_only_solar_thermal.py
@@ -48,7 +48,7 @@ def main():
         s.loadTopo()
 
         # # Create the distribution class
-        s.distribute['albedo'] = smrf.distribute.albedo.albedo(
+        s.distribute['albedo'] = smrf.distribute.albedo.Albedo(
             s.config['albedo'])
         s.distribute['solar'] = smrf.distribute.solar.solar(s.config['solar'],
                                                             s.distribute['albedo'].config,
